Move link click handling out of Markdown component
