+++ shared/src/components/Markdown.tsx
@@ -1,7 +1,7 @@
 import classNames from 'classnames'
 import React, { useCallback } from 'react'
 import * as H from 'history'
-import { isInstanceOf, anyOf } from '../util/types'
+import { createLinkClickHandler } from './linkClickHandler'
 
 interface Props {
     wrapper?: 'div' | 'span'
@@ -21,31 +21,7 @@ export const Markdown: React.FunctionComponent<Props> = ({
 }: Props) => {
     // Links in markdown cannot use react-router's <Link>.
     // Prevent hitting the backend (full page reloads) for links that stay inside the app.
-    const onClick = useCallback<React.MouseEventHandler<unknown>>(
-        event => {
-            // Do nothing if the link was requested to open in a new tab
-            if (event.ctrlKey || event.metaKey) {
-                return
-            }
-            // Check if click happened within an anchor inside the markdown
-            const anchor = event.nativeEvent
-                .composedPath()
-                .slice(0, event.nativeEvent.composedPath().indexOf(event.currentTarget))
-                .find(anyOf(isInstanceOf(HTMLAnchorElement), isInstanceOf(SVGAElement)))
-            if (!anchor) {
-                return
-            }
-            const href = typeof anchor.href === 'string' ? anchor.href : anchor.href.baseVal
-            // Check if URL is outside the app
-            if (!href.startsWith(window.location.origin)) {
-                return
-            }
-            // Handle navigation programmatically
-            event.preventDefault()
-            history.push(new URL(href).pathname)
-        },
-        [history]
-    )
+    const onClick = useCallback(createLinkClickHandler(history), [history])
     return (
         <RootComponent
             onClick={onClick}
